Refactor : 코드정리
