+++ src/main/java/com/sdy/bbbb/BbbbApplication.java
@@ -1,10 +1,8 @@
 package com.sdy.bbbb;
 
-import com.sdy.bbbb.data.DataTagName;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 import javax.annotation.PostConstruct;
