--- src/main/java/com/sdy/bbbb/data/dataDto/PopByHourDto.java
@@ -1,38 +1,0 @@
-package com.sdy.bbbb.data.dataDto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class PopByHourDto {
-    private Integer H0;
-    private Integer H1;
-    private Integer H2;
-    private Integer H3;
-    private Integer H4;
-    private Integer H5;
-    private Integer H6;
-    private Integer H7;
-    private Integer H8;
-    private Integer H9;
-    private Integer H10;
-    private Integer H11;
-    private Integer H12;
-    private Integer H13;
-    private Integer H14;
-    private Integer H15;
-    private Integer H16;
-    private Integer H17;
-    private Integer H18;
-    private Integer H19;
-    private Integer H20;
-    private Integer H21;
-    private Integer H22;
-    private Integer H23;
-
-}
