+++ src/main/java/com/sdy/bbbb/dto/response/data/BaseGuInfoDto.java
@@ -1,4 +1,4 @@
-package com.sdy.bbbb.data.dataDto;
+package com.sdy.bbbb.dto.response.data;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
