+++ src/main/java/com/sdy/bbbb/dto/response/data/SpotInfoDto.java
@@ -1,10 +1,7 @@
-package com.sdy.bbbb.data.dataDto;
+package com.sdy.bbbb.dto.response.data;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import java.util.List;
-import java.util.Map;
 
 @Getter
 @NoArgsConstructor
