--- src/main/java/com/sdy/bbbb/querydsl/HashTagRepositoryCustom.java
@@ -1,11 +1,0 @@
-package com.sdy.bbbb.querydsl;
-
-import com.sdy.bbbb.entity.HashTag;
-import com.sdy.bbbb.repository.HashTagRepository;
-
-import java.util.List;
-
-//public interface HashTagRepositoryCustom {
-////    List<HashTag> searchHotTag20();
-//    List<HashTag> searchHottest(String guName);
-//}
