+++ src/main/java/com/sdy/bbbb/s3/S3Uploader.java
@@ -1,77 +1,66 @@
-//package com.sdy.bbbb.s3;
-//
-//import com.amazonaws.services.s3.AmazonS3Client;
-//import com.amazonaws.services.s3.model.CannedAccessControlList;
-//import com.amazonaws.services.s3.model.PutObjectRequest;
-//import com.sdy.bbbb.exception.CustomException;
-//import com.sdy.bbbb.exception.ErrorCode;
-//import lombok.RequiredArgsConstructor;
-//import lombok.extern.slf4j.Slf4j;
-//import org.springframework.beans.factory.annotation.Value;
-//import org.springframework.stereotype.Component;
-//import org.springframework.web.multipart.MultipartFile;
-//
-//import java.io.File;
-//import java.io.FileOutputStream;
-//import java.io.IOException;
-//import java.util.Optional;
-//import java.util.UUID;
-//
-//@Component
-//@RequiredArgsConstructor
-//@Slf4j
-//public class S3Uploader {
-//    private final AmazonS3Client amazonS3Client;
-//
-//    @Value("${cloud.aws.s3.bucket}")
-//    private String bucket;
-//
-//    // 1. 클라이언트한테 받아온다.
-//    // 2. 파싱해서 s3에 저장후 그 url 디비에 저장.
-//    // 4. url 리턴
-//
-//    public String uploadFiles(MultipartFile multipartFile, String dirName) throws IOException {
-////        File file = new File(Objects.requireNonNull(multipartFile.getOriginalFilename()));
-////        multipartFile.transferTo(file); // 파일 변환할 수 없으면 에러
-//        File uploadFile = convert(multipartFile)
-//                .orElseThrow(() -> new CustomException(ErrorCode.FailConvertImage));
-//        return upload(uploadFile, dirName);
-//    }
-//
-//    public String upload(File uploadFile, String filePath) {
-//        String fileName = filePath + "/" + UUID.randomUUID() + uploadFile.getName();   // S3에 저장된 파일 이름
-//        String uploadImageUrl = putS3(uploadFile, fileName); // s3로 업로드
-//        removeNewFile(uploadFile);
-//        return uploadImageUrl;
-//    }
-//
-//    // S3로 업로드
-//    private String putS3(File uploadFile, String fileName) {
-//        amazonS3Client.putObject(new PutObjectRequest(bucket, fileName, uploadFile).withCannedAcl(CannedAccessControlList.PublicRead));
-//        return amazonS3Client.getUrl(bucket, fileName).toString();
-//    }
-//
-//    // 로컬에 저장된 이미지 지우기
-//    private void removeNewFile(File targetFile) {
-//        if (targetFile.delete()) {
-//            System.out.println("File delete success");
-//            return;
-//        }
-//        System.out.println("File delete fail");
-//    }
-//
-//    private Optional<File> convert(MultipartFile file) throws IOException {
-//        File convertFile = new File(System.getProperty("user.dir") + "/" + file.getOriginalFilename());
-//        if (convertFile.createNewFile()) { // 바로 위에서 지정한 경로에 File이 생성됨 (경로가 잘못되었다면 생성 불가능)
-//            try (FileOutputStream fos = new FileOutputStream(convertFile)) { // FileOutputStream 데이터를 파일에 바이트 스트림으로 저장하기 위함
-//                fos.write(file.getBytes());
-//            }
-//            return Optional.of(convertFile);
-//        }
-//        return Optional.empty();
-//    }
-//
-////    private void deleteS3Object(String fileName) {
-////        amazonS3Client.deleteObject(bucket, fileName);
-////    }
-//}
+package com.sdy.bbbb.s3;
+
+import com.amazonaws.AmazonServiceException;
+import com.amazonaws.SdkClientException;
+import com.amazonaws.services.s3.AmazonS3Client;
+import com.amazonaws.services.s3.model.CannedAccessControlList;
+import com.amazonaws.services.s3.model.DeleteObjectRequest;
+import com.amazonaws.services.s3.model.ObjectMetadata;
+import com.amazonaws.services.s3.model.PutObjectRequest;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.io.IOException;
+import java.util.UUID;
+
+@RequiredArgsConstructor
+@Component
+public class S3Uploader {
+
+    private final AmazonS3Client amazonS3Client;
+
+    @Value("${cloud.aws.s3.bucket}")
+    private String bucket;
+
+
+    public String upload(MultipartFile uploadFile, String dirName) {
+        String fileType = uploadFile.getOriginalFilename().substring(uploadFile.getOriginalFilename().lastIndexOf("."));
+        String randomName = UUID.randomUUID().toString() + fileType; // 파일 중복되지 않게 고유식별자 생성
+
+        String fileName = dirName + "/" + randomName;
+        String uploadImageUrl = putS3(uploadFile, fileName);
+        return uploadImageUrl;
+    }
+    public void deleteFileFromS3(String key) {
+        //key는 경로, 파일이름 풀로 ex) static/test.txt
+        deleteFile(key);
+    }
+
+    private String putS3(MultipartFile uploadFile, String fileName) {
+        ObjectMetadata metadata = new ObjectMetadata();
+        metadata.setContentType(uploadFile.getContentType());
+        metadata.setContentLength(uploadFile.getSize());
+        try {
+            amazonS3Client.putObject(new PutObjectRequest(bucket, fileName, uploadFile.getInputStream(), metadata).withCannedAcl(CannedAccessControlList.PublicRead));
+
+        }catch (IOException e) {
+            // TODO: handle exception
+            e.printStackTrace();
+        }
+        return amazonS3Client.getUrl(bucket, fileName).toString();
+    }
+
+    private void deleteFile(String key) {
+        DeleteObjectRequest deleteObjectRequest = new DeleteObjectRequest(bucket, key);
+        try {
+            amazonS3Client.deleteObject(deleteObjectRequest);
+        }catch(AmazonServiceException e) {
+            e.printStackTrace();
+        }catch (SdkClientException e) {
+            e.printStackTrace();
+        }
+    }
+
+}
